DataDriftTrigger: fix bugs, refactor
