+++ modyn/supervisor/internal/triggers/model_wrappers/__init__.py
@@ -3,10 +3,8 @@
 """
 import os
 
-from .datadrifttrigger import DataDriftTrigger  # noqa: F401
-from .amounttrigger import DataAmountTrigger  # noqa: F401
-from .timetrigger import TimeTrigger  # noqa: F401
-from .trigger import Trigger  # noqa: F401
+from .abstract_model_wrapper import AbstractModelWrapper  # noqa: F401
+from .modyn_model_wrapper import ModynModelWrapper  # noqa: F401
 
 files = os.listdir(os.path.dirname(__file__))
 files.remove("__init__.py")
